upd: finish modules
